+++ at_abc391_A.cpp
@@ -1,7 +1,7 @@
 #include <bits/stdc++.h>
 using namespace std;
 
-long long main(){
+int main(){
     string a;
     cin >> a;
     if(a=="N"){
