+++ at_abc393_B.cpp
@@ -1,13 +1,13 @@
 #include <bits/stdc++.h>
 using namespace std;
 
-long long main(){
+int main() {
     string s;
     cin >> s;
     long long anss = 0;
-    for (long long i = 0; i < s.size();i++){
-        for (long long p = 1; i + 2 * p < s.size();p++){
-            if(s[i]=='A' && s[i+p]=='B' && s[i+2*p]=='C'){
+    for (long long i = 0; i < s.size();i++) {
+        for (long long p = 1; i + 2 * p < s.size();p++) {
+            if (s[i] == 'A' && s[i + p] == 'B' && s[i + 2 * p] == 'C') {
                 anss++;
             }
         }
