+++ at_abc412_A.cpp
@@ -1,11 +1,11 @@
 #include <bits/stdc++.h>
 using namespace std;
 
-long long main(){
+int main() {
     long long n;
     cin >> n;
     long long anss = 0;
-    for (long long i = 0; i < n;i++){
+    for (long long i = 0; i < n;i++) {
         long long a, b;
         cin >> a >> b;
         anss += (b > a);
